requirements.txt
mysql-connector-python
python-dotenv
resend
customtkinter
